requirements.txt
whitenoise
Django==1.6.5
MySQL-python==1.2.5
Pillow==2.5.0
South==1.0
argparse==1.2.1
coverage==3.7.1
#django-debug-toolbar==1.2.1
django-extensions==1.3.8
django-filter==0.7
django-model-utils==2.0.3
django-nose==1.2
django-photologue==2.8
django-pipeline==1.3.25
django-registration==1.0
django-sortedm2m==0.8.0
django-twitter-bootstrap==3.1.1
djangorestframework==2.3.14
futures==2.1.6
ipython==2.1.0
jsonpickle==0.7.1
#mysql-connector-python==1.2.2
nose==1.3.3
numpy==1.8.1
pep8==1.5.7
#selenium==2.42.1
six==1.7.3
sqlparse==0.1.11
wsgiref==0.1.2
gunicorn
psycopg2
dj_database_url